Adding EPIC, CCDMinimizer, and minimization of pure pairwise energy
Adding the EPIC (Energy as Polynomials in Internal Coordinates)
algorithm, turned on using the useEPIC flag.  Also providing a faster
and more accurate minimizer, the CCDMinimizer (turned on by the useCCD
flag, which is on by default).  Previous versions of OSPREY minimized
the full energy of a rotamer pair when trying to get the minimum
pairwise energy, which is not provable though it results in higher
bounds than minimizing the pairwise energy alone; default behavior is
now to minimize the pairwise energy alone (the old heuristic can be
enabled with minimizePairwise=false).  Finally, the makeStericShell
top-level command has been added.

+++ src/AminoAcidTemplates.java
@@ -165,7 +165,7 @@ public class AminoAcidTemplates implements Serializable {
 				tmpName = getToken(curLine,2);
 				at.changeType(tmpName);
 				at.forceFieldType = getToken(curLine,3);
-				at.charge = (float) (new Float(getToken(curLine,11)).floatValue());
+				at.charge = (double) (new Double(getToken(curLine,11)).doubleValue());
 				at.isBBatom = at.setIsBBatom();
 				//KER: The first atom is bonded to a dummy atom so we can't include that
 				//KER: in the bond list, so check atom is >= 0
@@ -246,7 +246,7 @@ public class AminoAcidTemplates implements Serializable {
 						for(Atom a:r.atom){
 							if(a.name.equalsIgnoreCase(getToken(curLine,1))){
 								foundAtom = true;
-								a.setCoords(new Float(getToken(curLine,2)), new Float(getToken(curLine,3)), new Float(getToken(curLine,4)));
+								a.setCoords(new Double(getToken(curLine,2)), new Double(getToken(curLine,3)), new Double(getToken(curLine,4)));
 								break;
 							}
 						}
@@ -314,7 +314,7 @@ public class AminoAcidTemplates implements Serializable {
 				tmpName = getToken(curLine,2);
 				at.changeType(tmpName);
 				at.forceFieldType = getToken(curLine,3);
-				at.charge = (float) (new Float(getToken(curLine,11)).floatValue());
+				at.charge = (double) (new Double(getToken(curLine,11)).doubleValue());
 				at.isBBatom = at.setIsBBatom();
 				at.addBond(((new Integer(getToken(curLine,5))).intValue())-dumPresent);
 				newRes.addAtom(at);  // add atom
@@ -381,7 +381,7 @@ public class AminoAcidTemplates implements Serializable {
 				tmpName = getToken(curLine,2);
 				at.changeType(tmpName);
 				at.forceFieldType = getToken(curLine,3);
-				at.charge = (float) (new Float(getToken(curLine,11)).floatValue());
+				at.charge = (double) (new Double(getToken(curLine,11)).doubleValue());
 				at.isBBatom = at.setIsBBatom();
 				at.addBond(((new Integer(getToken(curLine,5))).intValue())-dumPresent);
 				newRes.addAtom(at);  // add atom
